+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.1.3: July 15th, 2020
 * Remove unused params
 * Remove regular price when on sale